requirements.txt
# Core Flask & Web
Flask==3.0.2
Flask-MySQLdb==2.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0

# Database
mysqlclient==2.2.1
bcrypt==4.1.2

# TensorFlow/Keras - UPDATED TO COMPATIBLE VERSIONS
tensorflow==2.15.0  # Last version with built-in Keras
numpy==1.26.4

# NLP
nltk==3.8.1

# Computer Vision
opencv-python==4.9.0.80
deepface==0.0.79
Pillow==10.2.0

# Gemini AI & Utilities
google-generativeai==0.3.2
python-dotenv==1.0.1
autocorrect==2.6.1